requirements.txt
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
future==0.18.2
Jinja2==2.11.2
jmespath==0.9.5
Js2Py==0.70
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
Naked==0.1.31
packaging==20.4
Pillow==7.1.1
pipenv==2018.11.26
pipwin==0.5.0
pyttsx3==2.90
pytz==2020.1
pywin32==228
SQLAlchemy==1.3.17
sqlparse==0.3.1

{
  "connectionMethod": "Server and Port",
  "mysqlOptions": {
    "authProtocol": "default",
    "enableSsl": "Disabled"
  },
  "previewLimit": 50,
  "server": "students",
  "port": 3306,
  "usePassword": "SQLTools Driver Credentials",
  "driver": "MySQL",
  "name": "students",
  "database": "students",
  "connectionTimeout": 30
}